pyproject.toml
[tool.poetry]
name = "fishsense_common"
version = "0.1.0"
description = ""
authors = ["Christopher L. Crutchfield <ccrutchf@ucsd.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
wakepy = "^0.10.1"
pyyaml = "^6.0.2"
ray = "^2.37.0"
torch = { version = "^2.4.0", source="torch"}
appdirs = "^1.4.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
black = "^24.3.0"
pylint = "^3.2.0"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu124"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
